safari-600.7-branch/LayoutTests/platform/ios-sim/fast/multicol/hit-test-gap-between-pages-flipped-expected.txt

CONSOLE MESSAGE: line 12: TypeError: null is not an object (evaluating 'document.caretRangeFromPoint(115, 550).startContainer')
FAIL: Test did not run (DumpRenderTree or WebKitTestRunner needed)
